+++ CHANGELOG.md
@@ -29,12 +29,19 @@ and this project adheres to [Semantic Versioning](https://semver.org).
 
 ## [0.0.0] - 2026-04-17
 
-#### Added
+### Added
 - `CHANGELOG.md`;
 - Authors section to `README.md`;
 
-#### Changed
+### Changed
 - `GUIDELINES.md` renamed to `CONTRIBUTING.md`;
 - Mention to `CONTRIBUTING.md` in `README.md`;
 
 ## Unreleased
+
+### Added
+- DomainException class;
+- Role domain class;
+- User domain class;
+- Course and CourseArtifact domain classes;
+- Assignment, AssignmentFlags and ForbiddenFlagCombination domain classes;